\name{bkgdMOCLocal}
\alias{bkgdMOCLocal}
\title{Makes the MOC background.}
\usage{
bkgdMOCLocal(df, gene_expr, SeuratObj, bkgd=5, assay=NULL, slot="data")
}
\description{
This is a nested function that should not be used on its own.}
\arguments{

}
\value{
Returns background information within the CoExpress() function.
}
\examples{
# N/A
}
