\name{bkgdMOCParition}
\alias{bkgdMOCParition}
\title{Makes the MOC background.}
\usage{
bkgdMOCParition(SeuratObj, gene_expr, permutations=5, paritions=4, assay=NULL, slot="data", top, bottom)
}
\description{
This is a nested function that should not be used on its own.}
\arguments{

}
\value{
Returns background information within the CoExpress() function.
}
\examples{
# N/A
}
