\name{GetCoExpr}
\alias{GetCoExpr}
\title{Create a dataframe with MOC information.}
\usage{
GetCoExpr(SeuratObj, geneA, geneB, assay=NULL, slot="data")
}
\description{
This is a nested function that should not be used on its own.}
\arguments{

}
\value{
Returns MOC information within the CoExpress() function.
}
\examples{
# N/A
}
