\name{Local_Once}
\alias{Local_Once}
\title{Does a local background query for one gene pair.}
\usage{
Local_Once(SOBJ, ass, sl, a, b, bkgd, gene_expr)
}
\description{
This is a nested function that should not be used on its own.}
\arguments{

}
\value{
Returns MOC background information within the CoExpress() function.
}
\examples{
# N/A
}
