\name{plotCoExpr}
\alias{plotCoExpr}
\title{Plots Co-expression clusters.}
\usage{
plotCoExpr(co)
}
\description{
Visualize co-expression clusters.}
\arguments{
\item{co}{\code{CoExpress()} output \code{data.frame}.}
}
\value{
Returns a grob, showing a clustered co-expression matrix of all the gene pairs tabulated in by \code{CoExpress()}.
}
\examples{
# plotCoExpr(CoExpr)
}
